Se comentarons los rests, se ajusto modelos y test

+++ CoachingSpring/src/main/java/pdg/dataaccess/api/Dao.java
@@ -40,5 +40,5 @@ public interface Dao<T, PK extends Serializable> {
 
     List<T> findByProperty(String propertyName, Object value);
 
-	public Double genSecuencia();
+	public long genSecuencia();
 }
